Level0 Shippable
Now uses correct exit prefab, adds background track, places used scripts
into level0 folder for tidyness.
