--- Assets/Scripts/BlokePath.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 74819b375cb90654ba8084b2636d5fe3
-timeCreated: 1460894099
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
