--- Assets/Scripts/FlipBlock.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: f982bc69689bf8449980d50662bcd3bf
-timeCreated: 1460827676
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
