+++ Assets/Scripts/Level0/CoinPop.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: b4f936683921ab940bc3201a19879a1e
+timeCreated: 1460896692
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
