--- Assets/Scripts/NpcPath.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 7582e90241832da49bac90965e02d780
-timeCreated: 1460824509
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
